+++ src-ui/app/views/dashboard-view/views/auto-invite-request-accept-view/auto-invite-request-accept-view.component.ts
@@ -1,6 +1,6 @@
 import { Component, DestroyRef, OnInit } from '@angular/core';
 import { SelectBoxItem } from '../../../../components/select-box/select-box.component';
-import { distinctUntilChanged } from 'rxjs';
+import { distinctUntilChanged, map, skip, tap } from 'rxjs';
 import { VRChatService } from '../../../../services/vrchat.service';
 import { hshrink, noop, vshrink } from '../../../../utils/animations';
 import { ModalService } from 'src-ui/app/services/modal.service';
@@ -17,7 +17,18 @@ import {
 import { AutomationConfigService } from '../../../../services/automation-config.service';
 import { ConfirmModalComponent } from '../../../../components/confirm-modal/confirm-modal.component';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import {
+  PlayerListPresetModalComponent,
+  PlayerListPresetModalInputModel,
+  PlayerListPresetModalOutputModel,
+} from '../../../../components/player-list-preset-modal/player-list-preset-modal.component';
+import { AppSettingsService } from '../../../../services/app-settings.service';
 
+interface PresetOptions {
+  onSleepEnable: SelectBoxItem;
+  onSleepDisable: SelectBoxItem;
+  onSleepPreparation: SelectBoxItem;
+}
 @Component({
   selector: 'app-auto-invite-request-accept-view',
   templateUrl: './auto-invite-request-accept-view.component.html',
@@ -50,11 +61,23 @@ export class AutoInviteRequestAcceptViewComponent implements OnInit {
   );
   isOnBusyStatus = false;
   isOnJoinMeStatus = false;
+  presetOptions: SelectBoxItem[] = [
+    {
+      id: 'NONE',
+      label: 'auto-invite-request-accept.options.presetOptions.options.NONE',
+    },
+  ];
+  presetOption: PresetOptions = {
+    onSleepEnable: this.presetOptions[0],
+    onSleepDisable: this.presetOptions[0],
+    onSleepPreparation: this.presetOptions[0],
+  };
 
   constructor(
     protected vrchat: VRChatService,
     private modal: ModalService,
     private automationConfig: AutomationConfigService,
+    private appSettings: AppSettingsService,
     private modalService: ModalService,
     private destroyRef: DestroyRef
   ) {}
@@ -70,6 +93,35 @@ export class AutoInviteRequestAcceptViewComponent implements OnInit {
       this.isOnBusyStatus = user?.status === 'busy';
       this.isOnJoinMeStatus = user?.status === 'join me';
     });
+    this.appSettings.settings
+      .pipe(
+        takeUntilDestroyed(this.destroyRef),
+        map((settings) => settings.playerListPresets),
+        distinctUntilChanged(),
+        tap((presets) => {
+          this.presetOptions = [
+            this.presetOptions[0],
+            ...presets.map((preset) => ({
+              id: preset.id,
+              label: preset.name,
+            })),
+          ];
+        }),
+        skip(1),
+        tap(() => {
+          // Reset options to NONE if the set preset could no longer be found
+          Object.keys(this.presetOption).forEach((key) => {
+            const keyT = key as keyof PresetOptions;
+            if (
+              this.presetOption[keyT].id !== 'NONE' &&
+              !this.presetOptions.find((o) => o.id === this.presetOption[keyT].id)
+            ) {
+              this.presetOption[keyT] = this.presetOptions[0];
+            }
+          });
+        })
+      )
+      .subscribe();
     this.automationConfig.configs
       .pipe(takeUntilDestroyed(this.destroyRef))
       .subscribe(async (configs) => {
@@ -79,6 +131,15 @@ export class AutoInviteRequestAcceptViewComponent implements OnInit {
           [...this.config.playerIds].sort(),
           this.playerList.map((o) => o.id).sort()
         );
+        this.presetOption['onSleepEnable'] =
+          this.presetOptions.find((o) => o.id === this.config.presetOnSleepEnable) ??
+          this.presetOptions[0];
+        this.presetOption['onSleepDisable'] =
+          this.presetOptions.find((o) => o.id === this.config.presetOnSleepDisable) ??
+          this.presetOptions[0];
+        this.presetOption['onSleepPreparation'] =
+          this.presetOptions.find((o) => o.id === this.config.presetOnSleepPreparation) ??
+          this.presetOptions[0];
         if (this.loggedIn && playersChanged) await this.refreshPlayerList();
       });
   }
@@ -154,8 +215,59 @@ export class AutoInviteRequestAcceptViewComponent implements OnInit {
       });
   }
 
+  async loadPreset() {
+    this.modalService
+      .addModal<PlayerListPresetModalInputModel, PlayerListPresetModalOutputModel>(
+        PlayerListPresetModalComponent,
+        {
+          mode: 'load',
+        },
+        {}
+      )
+      .subscribe((result) => {
+        if (result?.playerIds) {
+          this.updateConfig({ playerIds: result.playerIds });
+        }
+      });
+  }
+
+  async savePreset() {
+    this.modalService
+      .addModal<PlayerListPresetModalInputModel, PlayerListPresetModalOutputModel>(
+        PlayerListPresetModalComponent,
+        {
+          mode: 'save',
+          playerIds: this.playerList.map((p) => p.id),
+        }
+      )
+      .subscribe();
+  }
+
   async setListMode(id?: string) {
     if (!id || !this.listModeOptions.map((o) => o.id).includes(id)) return;
     await this.updateConfig({ listMode: id } as Partial<AutoAcceptInviteRequestsAutomationConfig>);
+  }
+
+  protected async setPresetOption(automation: keyof PresetOptions, presetId: string | 'NONE') {
+    let option = this.presetOptions.find((o) => o.id === presetId)!;
+    if (!option) option = this.presetOptions[0];
+    this.presetOption[automation] = option;
+    switch (automation) {
+      case 'onSleepEnable':
+        await this.updateConfig({
+          presetOnSleepEnable: presetId === 'NONE' ? null : presetId,
+        });
+        break;
+      case 'onSleepDisable':
+        await this.updateConfig({
+          presetOnSleepDisable: presetId === 'NONE' ? null : presetId,
+        });
+        break;
+      case 'onSleepPreparation':
+        await this.updateConfig({
+          presetOnSleepPreparation: presetId === 'NONE' ? null : presetId,
+        });
+        break;
+    }
   }
 }